+++ src/HomeWork/HW_28102020/Task_1/Plant.java
@@ -1,5 +1,7 @@
 package HomeWork.HW_28102020.Task_1;
 
+import java.util.Objects;
+
 public class Plant {
     private String plantType;
     private String plantName;
@@ -7,6 +9,20 @@ public class Plant {
     public Plant(String plantType, String plantName) {
         this.plantType = plantType;
         this.plantName = plantName;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Plant plant = (Plant) o;
+        return Objects.equals(plantType, plant.plantType) &&
+                Objects.equals(plantName, plant.plantName);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(plantType, plantName);
     }
 
     public String getPlantType() {
