+++ src/HomeWork/HW_28102020/Task_2/Persons.java
@@ -3,7 +3,7 @@ package HomeWork.HW_28102020.Task_2;
 import java.util.ArrayList;
 
 public class Persons {
-    String name;
+    private String name;
     private ArrayList<Products> arrayList;
 
     public Persons(String name, ArrayList<Products> arrayList) {
