+++ src/HomeWork/HW_28102020/Task_4/AutoApp.java
@@ -15,33 +15,33 @@ public class AutoApp {
         System.out.println("Если вы хотите найти авто по параметрам - введите 4");
         System.out.println("Если вы хотите завешить работу - введите 0");
         while (flag) {
-            int nn = Cars.inputChoiseInt();
+            int nn = AutoHouse.inputChoiseInt();
             switch (nn) {
                 case 1:
-                    autoSaloon.add(Cars.addCars());
+                    autoSaloon.add(AutoHouse.addCars());
                     System.out.println("Авто успешно добавлено.");
                     System.out.println("Сделайте выбор:");
                     break;
                 case 2:
                     if(autoSaloon.size()==0){
-                        new AutoException().isEmpty();
+                        new AutoException().isEmpty("Автомобилей в салоне нет!");
                     }else{
                         for (Cars ca : autoSaloon) {
                             System.out.println(ca.toString());
                         }
                         System.out.println("Введите номер кузова авто для продажи:");
-                        int numCars = Cars.inputChoiseInt();
-                        if (Cars.soledCars(autoSaloon, numCars)) {
+                        int numCars = AutoHouse.inputChoiseInt();
+                        if (AutoHouse.soledCars(autoSaloon, numCars)) {
                             System.out.println("Автомобиль успешно продан");
                         } else {
-                            new AutoException().isAvailable();
+                            new AutoException().isAvailable("Такого автомобиля нет в автосалоне! ");
                         }
                     }
                     System.out.println("Сделайте выбор:");
                     break;
                 case 3:
                     if(autoSaloon.size()==0){
-                        new AutoException().isEmpty();
+                        new AutoException().isEmpty("Автомобилей в салоне нет!");
                     }else{
                         for (Cars ca : autoSaloon) {
                             System.out.println(ca.toString());
@@ -52,14 +52,14 @@ public class AutoApp {
                 case 4:
                     System.out.println("Укажите параметры поиска, если парамерт не важен введите 0");
                     System.out.println("Марка автомобиля:");
-                    String markAuto=Cars.inputChoiseString();
+                    String markAuto=AutoHouse.inputChoiseString();
                     System.out.println("Цвет автомобиля:");
-                    String colorAuto=Cars.inputChoiseString();
+                    String colorAuto=AutoHouse.inputChoiseString();
                     System.out.println("Максимальная стоимость автомобиля:");
-                    int priceAuto=Cars.inputChoiseInt();
-                    int quantityAuto=Cars.findAuto(autoSaloon, markAuto, colorAuto, priceAuto);
+                    int priceAuto=AutoHouse.inputChoiseInt();
+                    int quantityAuto=AutoHouse.findAuto(autoSaloon, markAuto, colorAuto, priceAuto);
                     if(quantityAuto==0){
-                        new AutoException().isAvailable();
+                        new AutoException().isAvailable("Такого автомобиля нет в автосалоне! ");
                     }else{
                         System.out.printf("Всего найдено %d автомобилей \n", quantityAuto);
                     }
