+++ src/HomeWork/HW_28102020/Task_4/AutoException.java
@@ -1,8 +1,8 @@
 package HomeWork.HW_28102020.Task_4;
 
 public class AutoException extends Exception{
-    public void isAvailable(){
-        System.out.println("Такого автомобиля нет в автосалоне! ");
+    public void isAvailable(String message){
+        System.out.println(message);
     }
-    public void isEmpty(){System.out.println("Автомобилей в салоне нет!");}
+    public void isEmpty(String message){System.out.println(message);}
 }
